Version 1.2.0

+++ src/main/java/org/trivait/minigamesmod/leaderboard/LeaderboardEntry.java
@@ -2,5 +2,5 @@ package org.trivait.minigamesmod.leaderboard;
 
 public record LeaderboardEntry(
         String name,
-        int value
+        long value
 ) {}
